requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
newrelic==7.0.0.166
pytz==2021.3
sqlparse==0.4.2
